vm: fix some size_t/uint64_t issues (5)

+++ src/cshared/buffer.cpp
@@ -4,7 +4,7 @@ Buffer::Buffer()
 {
 }
 
-Buffer::Buffer(const Buffer &other, size_t start, size_t length)
+Buffer::Buffer(const Buffer &other, uint64_t start, uint64_t length)
 {
     this->data         = other.data + start;
     this->length       = length;
@@ -41,7 +41,7 @@ Buffer &Buffer::operator=(const Buffer &other)
     return *this;
 }
 
-void Buffer::resize(size_t new_length)
+void Buffer::resize(uint64_t new_length)
 {
     byte *new_data = (byte *)malloc(new_length);
     if (data)
@@ -53,10 +53,10 @@ void Buffer::resize(size_t new_length)
     this->length = new_length;
 }
 
-void Buffer::append(const byte *data, size_t data_length)
+void Buffer::append(const byte *data, uint64_t data_length)
 {
-    const size_t old_length = this->length;
-    const size_t new_length = old_length + data_length;
+    const uint64_t old_length = this->length;
+    const uint64_t new_length = old_length + data_length;
     assert(new_length > old_length);
     resize(new_length);
     memcpy(this->data + old_length, data, data_length);
